refactor(api): deduplicate DSL shared entities into dsl_entities.py
