+++ api/tests/unit_tests/services/test_app_dsl_service_import_yaml_url.py
@@ -4,7 +4,8 @@ import httpx
 
 from models import Account
 from services import app_dsl_service
-from services.app_dsl_service import AppDslService, ImportMode, ImportStatus
+from services.app_dsl_service import AppDslService
+from services.entities.dsl_entities import ImportMode, ImportStatus
 
 
 def _build_response(url: str, status_code: int, content: bytes = b"") -> httpx.Response:
